Add an example with mobile phones. Correction of a bug during the projection when there is dictionary entries on the form x0 = 0 + .. with x0 the variable added during the first phase. Add the support of doubles

+++ src/fr/enslyon/DivisionRing/RationalDivisionRing.java
@@ -32,8 +32,8 @@ public class RationalDivisionRing implements DivisionRing<RationalNumber> {
 
     @Override
     public int compare(RationalNumber x, RationalNumber y) {
-        int a = x.getNominator() * y.getDenominator();
-        int b = x.getDenominator() * y.getNominator();
+        long a = x.getNominator() * y.getDenominator();
+        long b = x.getDenominator() * y.getNominator();
 
         if(a < b)
             return -1;
